Improved code for Payment status page and added different component for TreesforJane tenant
