+++ styles/donations.scss
@@ -523,15 +523,14 @@
   align-self: center;
   position: relative;
   display: flex;
-  justify-content: center;
+  justify-content: space-between;
+  flex-direction: column;
 }
 
 .donation-count {
   font-size: 16px;
   font-weight: 600;
   text-align: center;
-  position: absolute;
-  bottom: 20px;
   color: $light;
   font-variant-ligatures: none;
 }
@@ -544,6 +543,7 @@
   margin-top: 16px;
   color: $light;
   font-variant-ligatures: none;
+  margin-bottom: 40px;
 }
 
 .thankyou-image-header {
@@ -554,6 +554,7 @@
   font-variant-ligatures: none;
   padding-top: 15%;
   white-space: pre-line;
+  flex-grow: 1;
 }
 
 .thankyou-image-header p h1 {
@@ -573,14 +574,13 @@
   }
   .donation-count {
     font-size: 130px;
-    bottom: 300px;
     padding: 0px 80px;
+    margin-bottom: 200px;
   }
   .donation-url {
     font-size: 75px;
     margin: 0px;
-    position: absolute;
-    bottom: 60px;
+    margin-bottom: 100px;
   }
   .thankyou-image-header {
     font-size: 100px;
